@@ -33,9 +33,14 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.16.2",
-    "react": "^15.4.2",
+    "bootstrap": "^3.3.7",
+    "jquery": "^3.2.1",
+    "json": "^9.0.6",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
     "react-bootstrap": "^0.30.7",
-    "react-dom": "^15.4.2",
+    "react-bootstrap-slider": "^2.0.0",
+    "react-dom": "^15.6.1",
     "react-redux": "^5.0.2",
     "react-router": "^3.0.1",
     "react-router-bootstrap": "^0.23.1",
